SPAZA-228 Remove Target.apparent_radec
It was always an iffy function and since the parallactic angle doesn't
need it anymore, it is safe to remove. If you need apparent (ra, dec),
rather use Astropy CIRS (or a similar coordinate frame) directly.

Also deprecate `Target.astrometric_radec`, now that there is only one
version of radec. Issue a FutureWarning for folks still using it.